Fetch-lib + rename to cos-agent
